requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
sqlalchemy==2.0.35
psycopg[binary]>=3.2.2
groq==0.11.0
apscheduler==3.10.4
httpx==0.27.0
pydantic-settings==2.5.0
